+++ tests/test_model_api.py
@@ -170,7 +170,10 @@ def _create_model(model, tmpdir):
 
     # results
     results = _create_dynmaps(model, mod.staticmaps, forcing=False)
-    mod.set_results(results)
+    with pytest.raises(ValueError):
+        mod.set_results(results)
+    mod.set_results(results, name="results")
+    mod.set_results(results, split_dataset=True)
     testres = results["waterlevel"].rename("test")
     with pytest.raises(ValueError):
         mod.set_results(testres.values)
@@ -263,6 +266,12 @@ def test_model_method(model, tmpdir):
     with pytest.raises(ValueError):
         # model_out None with mode=r
         mod1.update(opt=opt)
+    with pytest.raises(IOError):
+        mod1.write_staticmaps()
+    with pytest.raises(IOError):
+        mod1.write_staticgeoms()
+    with pytest.raises(IOError):
+        mod1.write_config()
     opt = {
         "setup_config": {"global.name": "test"},
         "setup_basemaps": {"add_geom": True},
@@ -321,8 +330,6 @@ def test_object(model, tmpdir):
             mod1.write_forcing()
         with pytest.raises(IOError):
             mod1.write_states()
-        with pytest.raises(IOError):
-            mod1.write_results()
         mod1 = TESTMODELS.get(model)(root=root, mode="w")
         with pytest.raises(NotImplementedError):
             mod1.write_staticmaps()
@@ -332,8 +339,6 @@ def test_object(model, tmpdir):
             mod1.write_forcing()
         with pytest.raises(NotImplementedError):
             mod1.write_states()
-        with pytest.raises(NotImplementedError):
-            mod1.write_results()
 
 
 p.stop()